requirements.txt
dlib==20.0.0
face-recognition==1.3.0
numpy>=1.24.2
Pillow>=9.4.0

